pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atlas-fabric"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-anthropic",
    "langchain-azure-ai",
    "langchain-openai",
    "langchain-google-genai",
    "langchain-ollama",
    "langchain-groq",
    "langchain-mistralai",
    "shapely>=2.0",
    "geopandas",
    "motor>=3.6",
    "fastapi>=0.115",
    "uvicorn[standard]",
    "typer>=0.13",
    "pydantic>=2.0",
    "opentelemetry-sdk>=1.20.0",
    "openinference-instrumentation-langchain>=0.1.0",
    "pymongo>=4.0.0",
    "slowapi",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "httpx",
    "bandit",
    "pip-audit",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
follow_imports = "skip"

[[tool.mypy.overrides]]
module = ["dotenv", "dotenv.*", "certifi", "certifi.*"]
follow_imports = "skip"
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["agents", "api", "cli", "evals", "geo", "knowledge", "storage"]

[tool.pytest.ini_options]
testpaths = ["tests"]
